Projet finished

+++ controllers/sauces.js
@@ -1,8 +1,11 @@
+// CREATION DU MIDDLEWARE POUR LA GESTION DES SAUCES
+
 const Sauce = require("../models/sauce");
 const fs = require("fs");
 const { findOneAndDelete } = require("../models/sauce");
 const { default: cluster } = require("cluster");
 
+// Récupération de toutes les sauces
 exports.getAllSauces = (req, res, next) => {
   Sauce.find()
     .then((sauces) => {
@@ -13,10 +16,9 @@ exports.getAllSauces = (req, res, next) => {
         error: error,
       });
     });
-
-  //res.status(200).json({ message: "Récupérer toutes les sauces" });
 };
 
+// Récupération d'une sauce par son identifiant
 exports.getById = (req, res, next) => {
   Sauce.findOne({
     _id: req.params.id,
@@ -31,17 +33,21 @@ exports.getById = (req, res, next) => {
     });
 };
 
+// Publication d'une sauce
 exports.postPictureAndSauce = (req, res, next) => {
   const sauceObject = JSON.parse(req.body.sauce);
   delete sauceObject._id;
   delete sauceObject._userId;
   const sauce = new Sauce({
+    // ... = spread : prend tout ce qu'il y a dedans
     ...sauceObject,
     userId: req.token.userId,
-    imageUrl: `${req.protocol}://${req.get("host")}/images/${
-      req.file.filename
-    }`,
+    imageUrl: "",
   });
+  if (req.file)
+    sauce.imageUrl = `${req.protocol}://${req.get("host")}/images/${
+      req.file.filename
+    }`;
   sauce
     .save()
     .then(() => {
@@ -52,6 +58,7 @@ exports.postPictureAndSauce = (req, res, next) => {
     });
 };
 
+// Modification de la sauce
 exports.putUpdateSauce = (req, res, next) => {
   const sauceObject = req.file
     ? {
@@ -61,27 +68,29 @@ exports.putUpdateSauce = (req, res, next) => {
         }`,
       }
     : { ...req.body };
-
-  delete sauceObject._userId;
   Sauce.findOne({ _id: req.params.id })
     .then((sauce) => {
       if (sauce.userId != req.token.userId) {
-        // req.auth auparavant
         res.status(401).json({ message: "Not authorized" });
       } else {
-        ////////////////// CODE RAJOUTE ///////////////////
         const filename = sauce.imageUrl.split("/images/")[1];
-        fs.unlink(`images/${filename}`, () => {
-          Sauce.deleteOne({ _id: req.params.id });
-        });
-        ///////////////////////////////////////////////////
-
-        Sauce.updateOne(
-          { _id: req.params.id },
-          { ...sauceObject, _id: req.params.id }
-        )
-          .then(() => res.status(200).json({ message: "Objet modifié!" }))
-          .catch((error) => res.status(401).json({ error }));
+        if (filename !== "") {
+          fs.unlink(`images/${filename}`, () => {
+            Sauce.updateOne(
+              { _id: req.params.id },
+              { ...sauceObject, _id: req.params.id }
+            )
+              .then(() => res.status(200).json({ message: "Objet modifié!" }))
+              .catch((error) => res.status(401).json({ error }));
+          });
+        } else {
+          Sauce.updateOne(
+            { _id: req.params.id },
+            { ...sauceObject, _id: req.params.id }
+          )
+            .then(() => res.status(200).json({ message: "Objet modifié!" }))
+            .catch((error) => res.status(401).json({ error }));
+        }
       }
     })
     .catch((error) => {
@@ -89,6 +98,7 @@ exports.putUpdateSauce = (req, res, next) => {
     });
 };
 
+// Suppression d'une sauce
 exports.deleteById = (req, res, next) => {
   Sauce.findOne({ _id: req.params.id })
     .then((sauce) => {
@@ -96,11 +106,12 @@ exports.deleteById = (req, res, next) => {
         res.status(401).json({ message: "Not authorized" });
       } else {
         const filename = sauce.imageUrl.split("/images/")[1];
-        fs.unlink(`images/${filename}`, () => {
-          Sauce.deleteOne({ _id: req.params.id })
-            .then(() => res.status(200).json({ message: "Objet supprimé !" }))
-            .catch((error) => res.status(401).json({ error }));
-        });
+        if (filename !== "")
+          fs.unlink(`images/${filename}`, () => {
+            Sauce.deleteOne({ _id: req.params.id })
+              .then(() => res.status(200).json({ message: "Objet supprimé !" }))
+              .catch((error) => res.status(401).json({ error }));
+          });
       }
     })
     .catch((error) => {
@@ -108,40 +119,29 @@ exports.deleteById = (req, res, next) => {
     });
 };
 
+// Likes et Dislikes
 exports.postLike = (req, res, next) => {
-  // 1 récupérer la sauce par son ID
-  // 2 si utilisateur à aimé la sauce, on ajoute 1 dans likes et son userId dans UserLike
-  // 3 si utilisateur n'a pas aimé la sauce, on ajoute 1 dans dislikes et son UserId dans UserDislikes
-  // 4 si il retire sa réaction
-  // 4.1 = si userId est dans tableau UserLikes on retire le like
-  // 4.2 = sinon on retire le dislike
+  // Fonctionnement
+  // 1 = On récupère la sauce par son ID
+  // 2 = Si utilisateur à aimé la sauce, on ajoute 1 dans likes et son userId dans UserLike
+  // 3 = Si utilisateur n'a pas aimé la sauce, on ajoute 1 dans dislikes et son UserId dans UserDislikes
+  // 4 = Si il retire sa réaction
+  // 4.1 = Si userId est dans tableau UserLikes on retire le like
+  // 4.2 = Sinon on retire le dislike
 
   Sauce.findOne({
     _id: req.params.id,
   }).then((sauce) => {
-    //////////////////////////////
-
     if (req.body.like == 1) {
-      // On vérifie si l'usersId est dans le tableau usersLiked
-      // si oui = message erreur
-      // si non = l'exécution du code continue
       if (sauce.usersLiked.includes(req.token.userId)) {
         return res.status(400).json({ message: "Sauce déjà likée" });
-      }
-      // On vérifie si l'usersId n'est pas déjà dans le tableau usersLiked, si il n'y est pas, on le
-      // rajoute dans le tableau usersLiked et on incrémente le compteur like
-      else if (sauce.usersDisliked.includes(req.token.userId)) {
+      } else if (sauce.usersDisliked.includes(req.token.userId)) {
         return res.status(400).json({ message: "Sauce déjà dislikée" });
       } else {
         sauce.likes++;
         sauce.usersLiked.push(req.token.userId);
       }
-
-      /////////////////////////////////////////////////////////////////////
     } else if (req.body.like == 0) {
-      // On vérifie si l'utilisateur n'avait ni liké, ni disLiké sa sauce (si son userId est présent
-      // dans usersLiked ou usersDisliked). Si oui, on décrémente le compteur et on retire son userId
-      // de usersLiked ou usersDisliked.
       if (
         !sauce.usersDisliked.includes(req.token.userId) &&
         !sauce.usersLiked.includes(req.token.userId)
@@ -150,24 +150,17 @@ exports.postLike = (req, res, next) => {
       } else if (sauce.usersDisliked.includes(req.token.userId)) {
         sauce.dislikes--;
         sauce.usersDisliked.splice(
-          sauce.usersDisliked.indexOf(req.token.userId),
-          1
+          sauce.usersDisliked.indexOf(req.token.userId)
         );
       } else {
-        // Si l'userId est dans les likes
         sauce.likes--;
-        sauce.usersLiked.splice(sauce.usersLiked.indexOf(req.token.userId), 1); ////////////////// ????????
+        sauce.usersLiked.splice(sauce.usersLiked.indexOf(req.token.userId), 1);
       }
     } else if (req.body.like == -1) {
-      // On vérifie si l'usersId est dans le tableau usersLiked
-      // si oui = message erreur
-      // si non = l'exécution du code continue
       if (sauce.usersLiked.includes(req.token.userId)) {
         return res.status(400).json({ message: "Sauce déjà likée" });
       } else if (sauce.usersDisliked.includes(req.token.userId)) {
         return res.status(400).json({ message: "Sauce déjà dislikée" });
-        // On vérifie si l'usersId n'est pas déjà dans le tableau usersDisliked, si il n'y est pas, on le
-        // rajoute dans le tableau usersLiked et on incrémente le compteur like
       } else {
         sauce.dislikes++;
         sauce.usersDisliked.push(req.token.userId);
